uncessary classes removed from ImageModel.js
