+++ vue-project/src/models/image/ImageDimensionsErrorMessage.js
@@ -1,0 +1,16 @@
+export class Errors {
+    /**
+     * @type {string}
+     */
+    width;
+
+    /**
+     * @type {string}
+     */
+    height;
+
+    constructor() {
+        this.width = '';
+        this.height = '';
+    }
+}
